+++ commands/weather.js
@@ -1,12 +1,23 @@
 const { SlashCommandBuilder } = require('discord.js');
 const { weather_token, weather_url } = require('../config.json');
+const request = require('request');
+const xml2js = require('xml2js');
+
+const parseXML = (xml) => {
+    return new Promise((resolve, reject) => {
+        xml2js.parseString(xml, (err, result) => {
+            if (err) reject(err);
+            else resolve(result);
+        });
+    });
+};
 
 module.exports = {
     data: new SlashCommandBuilder()
         .setName('날씨') // '/ping'으로 명령어 이름 지정 (한글 가능, 대문자 불가능)
         .setDescription('현재 서울 날씨 알림'), // '/ping'을 대화 채널에 타이핑 했을 때 뜨는 도움말
     async execute(interaction) {
-
+        // 현재 날짜와 시간 (예: 20210325, 0600)
         const now = new Date();
         const base_date = now.toISOString().split('T')[0].replace(/-/g, ''); // YYYYMMDD 형식
         const base_time = `${now.getHours()}00`; // 시간은 HH00 형식
